AI_chatbot/backend/requirements.txt
aiohttp==3.11.16
aiosignal==1.3.2
anyio==4.9.0
async-timeout==5.0.1
attrs==25.3.0
boto3==1.37.33
botocore==1.37.33
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
fastapi==0.115.12
frozenlist==1.5.0
h11==0.14.0
httptools==0.6.4
idna==3.10
jmespath==1.0.1
multidict==6.4.3
openai==1.72.0
pydantic==2.11.3
python-dateutil==2.9.0
python-decouple==3.8
python-dotenv==1.1.0
python-multipart==0.0.20
PyYAML==6.0.2
requests==2.32.4
s3transfer==0.11.4
six==1.17.0
sniffio==1.3.1
starlette==0.46.0
tqdm==4.67.1
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.5
websockets==15.0.1
yarl==1.19.0